remove deprecated imports. fetch real questions from cloud func.
