+++ src/app/feature_modules/quiz/components/phase/phase-layout/phase-layout.component.ts
@@ -1,22 +1,22 @@
-import { ThrowStmt } from '@angular/compiler'
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core'
-import { Box, MultipleChoicesProvider, ShortAnswerProvider } from '../../../models'
-import { Question } from '../../../models/interfaces'
+import { MultipleChoicesProvider, ShortAnswerProvider } from '../../../models'
 
 @Component({
   selector: 'app-phase-layout',
   templateUrl: './phase-layout.component.html',
-  styleUrls: ['./phase-layout.component.scss'],
+  styleUrls: ['./phase-layout.component.scss']
 })
 export class PhaseLayoutComponent implements OnInit {
   @Input() openQuestion: string
+
   @Input() phaseQuestions: MultipleChoicesProvider | ShortAnswerProvider
+
+  // tslint:disable-next-line:no-output-on-prefix
   @Output() onSubmit = new EventEmitter()
-  constructor() { }
 
   ngOnInit(): void {
     // console.log("This is phase Layout")
-    console.log(this.phaseQuestions);
+    // console.log(this.phaseQuestions)
   }
 
   submit(answear) {
