+++ src/app/feature_modules/quiz/containers/quiz-app/quiz-app.component.ts
@@ -1,27 +1,27 @@
 import { Component, ElementRef, Inject, Input, ViewChild } from '@angular/core'
 import { QuizPlayable } from '../../models'
 import { QuizHandler, QUIZ_STATE } from '../../services'
-import { realDummyData } from '../../models/dummy_data'
+
 @Component({
   selector: 'app-quiz',
   templateUrl: './quiz-app.component.html',
   providers: [
     {
       provide: QUIZ_STATE,
-      useClass: QuizHandler,
-    },
-  ],
+      useClass: QuizHandler
+    }
+  ]
 })
 export class QuizAppComponent {
-  fakeData = realDummyData['GID11060001']
   // answearType: QuestionType
   @ViewChild('helpQuestionRef') quizHelpBox: ElementRef<HTMLElement>
 
-  enableHelp = true;
+  enableHelp = true
+
   @Input()
   set quizId(value: string) {
     this.service.qid = value
-    console.log(this.service.qid)
+    // console.log(this.service.qid)
   }
 
   get hint(): string {
@@ -30,23 +30,23 @@ export class QuizAppComponent {
   }
 
   constructor(@Inject(QUIZ_STATE) public service: QuizPlayable) {
-    //TODO core feature calling to services
+    // TODO core feature calling to services
     // INPUT OUTPUT TO CHILD
   }
 
   toggleHelp() {
-    console.log(this.service)
+    // console.log(this.service)
     this.enableHelp = !this.enableHelp
     setTimeout(() => {
       this.quizHelpBox.nativeElement.scrollIntoView({
         behavior: 'smooth',
-        block: 'start',
+        block: 'start'
       })
     }, 100)
   }
 
   checkAnswear(event) {
     this.service.mainQuestion.submit(event)
-    console.log(this.service.mainQuestion.isCorrect)
+    // console.log(this.service.mainQuestion.isCorrect)
   }
 }
