+++ src/app/feature_modules/quiz/containers/quiz-phase-wrapper/quiz-phase-wrapper.component.ts
@@ -1,25 +1,17 @@
-import { ThrowStmt } from '@angular/compiler'
-import { Component, Input, OnInit } from '@angular/core'
-import { PossibleInputAnswer, Question } from '../../models'
-import {
-  MultiPhasesProvider,
-  MultiBoxesProvider,
-} from '../../models'
-import { realDummyData } from '../../models/dummy_data'
-import { PhaseStack } from '../../models/interfaces'
+import { Component, Input } from '@angular/core'
+import { PossibleInputAnswer } from '../../models'
+import { MultiPhasesProvider } from '../../models'
 
 @Component({
   // Render [subQuestions] xN
   selector: 'app-quiz-phase-wrapper',
   templateUrl: './quiz-phase-wrapper.component.html',
-  styleUrls: ['./quiz-phase-wrapper.component.scss'],
+  styleUrls: ['./quiz-phase-wrapper.component.scss']
 })
-export class QuizPhaseWrapperComponent implements OnInit {
+export class QuizPhaseWrapperComponent {
   @Input() phaseProvider: MultiPhasesProvider
-  // @Input() multiQuestionProvider: MultiQuestionsProvider
-  constructor() { }
 
-  ngOnInit(): void { }
+  // @Input() multiQuestionProvider: MultiQuestionsProvider
   getPhaseList() {
     return Array(this.phaseProvider.count).fill(0)
   }
