+++ src/app/feature_modules/quiz/models/index.ts
@@ -1,7 +1,6 @@
 export { MultiPhasesProvider } from './multi_phases_provider.class'
 export { MultiBoxesProvider } from './multi_boxes_provider.class'
 export { QuestionProvider, ShortAnswerProvider, MultipleChoicesProvider } from './question_provider.class'
-export { BoxType } from './enums.new'
-export { QuizState } from './enums'
-export * from './dummy_data.new'
-export * from './interfaces.new'
+export { BoxType, QuizState } from './enums'
+export * from './dummy_data'
+export * from './interfaces'
