+++ src/app/feature_modules/quiz/models/multi_phases_provider.class.ts
@@ -1,4 +1,4 @@
-import { Box, Phase, PossibleInputAnswer, Submitable } from './interfaces.new'
+import { Box, Phase, PossibleInputAnswer, Submitable } from './interfaces'
 import { MultiBoxesProvider } from './multi_boxes_provider.class'
 import { isNotUndefined } from '../../../shared'
