+++ src/app/feature_modules/quiz/models/question_provider.class.ts
@@ -1,5 +1,5 @@
-import { Box, Choice, MultipleChoices, PossibleInputAnswer, Question, ShortAnswer, Submitable } from './interfaces.new'
-import { BoxType } from './enums.new'
+import { Box, Choice, MultipleChoices, PossibleInputAnswer, Question, ShortAnswer, Submitable } from './interfaces'
+import { BoxType } from './enums'
 import { isNotUndefined, isNullOrUndefined, isUndefined } from '../../../shared'
 
 // Acts as substitution to `QuestionProvider`
@@ -74,7 +74,7 @@ export class ShortAnswerProvider extends QuestionProvider implements Submitable<
   private _submission?: number | string
 
   get correctAnswear(): string | number {
-    return this._question.answer;
+    return this._question.answer
   }
   get submission(): string | undefined {
     return isUndefined(this._submission) ? undefined : `${this._submission}`
