+++ src/app/feature_modules/quiz/quiz.module.ts
@@ -4,25 +4,21 @@ import { PhaseIndicatorComponent } from './components/phase/phase-indicator/phas
 import { QuizHeaderComponent } from './components/quiz-header/quiz-header.component'
 import { KatexModule } from 'ng-katex'
 import { QuizAppComponent } from './containers/quiz-app/quiz-app.component'
-import {
-  QUIZ_STATE,
-  QuizHandler,
-  QuizStorage,
-} from './services/quiz-provider.service'
 import { AnswerCardComponent } from './containers/answer-card.component'
 import { QuizShowcaseComponent } from './components/quiz-showcase.component'
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms'
 import { AnswerListViewComponent } from './components/answers-type/list/answer-type-list.component'
-import { AnswerTextViewComponent } from "./components/answers-type/input-box/answer-type-input-box.component";
-import { QuizPhaseWrapperComponent } from './containers/quiz-phase-wrapper/quiz-phase-wrapper.component';
-import { PhaseLayoutComponent } from './components/phase/phase-layout/phase-layout.component';
-import { AnswearTypeComponent } from './components/answers-type/answer-type.componet';
-import { QuizMainQuestionComponent } from './containers/quiz-main-question/quiz-main-question.component';
+import { AnswerTextViewComponent } from './components/answers-type/input-box/answer-type-input-box.component'
+import { QuizPhaseWrapperComponent } from './containers/quiz-phase-wrapper/quiz-phase-wrapper.component'
+import { PhaseLayoutComponent } from './components/phase/phase-layout/phase-layout.component'
+import { AnswearTypeComponent } from './components/answers-type/answer-type.componet'
+import { QuizMainQuestionComponent } from './containers/quiz-main-question/quiz-main-question.component'
 import { QuizWikiComponent } from './components/quiz-wiki/quiz-wiki.component'
 import { SharedModule } from 'src/app/shared'
+import { HttpClientModule } from '@angular/common/http'
 
 @NgModule({
-  imports: [CommonModule, FormsModule, ReactiveFormsModule, KatexModule, SharedModule],
+  imports: [CommonModule, FormsModule, ReactiveFormsModule, KatexModule, SharedModule, HttpClientModule],
   exports: [QuizAppComponent],
   declarations: [
     // Containers
@@ -39,16 +35,10 @@ import { SharedModule } from 'src/app/shared'
     PhaseLayoutComponent,
     AnswearTypeComponent,
     QuizMainQuestionComponent,
-    QuizWikiComponent,
+    QuizWikiComponent
 
     // Attribute Directives
-  ],
-  providers: [
-    QuizStorage,
-    {
-      provide: QUIZ_STATE,
-      useClass: QuizHandler,
-    },
-  ],
+  ]
 })
-export class QuizModule { }
+export class QuizModule {
+}
